% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcmdata-package.R
\docType{package}
\name{dcmdata-package}
\alias{dcmdata}
\alias{dcmdata-package}
\title{dcmdata: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line). Continuation lines should be indented.
}
\author{
\strong{Maintainer}: W. Jake Thompson \email{wjakethompson@gmail.com} (\href{https://orcid.org/0000-0001-7339-0300}{ORCID})

}
\keyword{internal}
