% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pie.R
\docType{data}
\name{pie_ft_data}
\alias{pie_ft_data}
\alias{pie}
\alias{pie_ft_qmatrix}
\alias{pie_pilot_data}
\alias{pie_pilot_qmatrix}
\title{Pathways for instructionally embedded assessment (PIE)}
\format{
\code{pie_ft_data} is a \link[tibble:tibble-package]{tibble} containing PIE
response data from the initial field test with 172 rows and 16 variables.
\itemize{
\item \code{student}: Respondent identifier.
\item \code{00592}-\code{88063}: Dichotomous item responses to the 16 PIE items for the
5.RA.A.1b learning pathway.
}

\code{pie_ft_qmatrix} is a \link[tibble:tibble-package]{tibble} that
identifies which levels of the learning pathway are measured by each PIE
item. The PIE field test utilized 15 items to measure the 3 levels in the
5.RA.A.1b pathway. The \code{pie_ft_qmatrix} correspondingly is
made up of 15 rows and 4 variables.
\itemize{
\item \code{task}: Item identifier, corresponds to \code{00592}-\code{88063} in
\code{pie_ft_data}.
\item \code{L1}, \code{L2}, and \code{L2}: Dichotomous indicator for whether or not the level is
measured by each item. A value of \code{1} indicates the level is measured by
the item and a value of \code{0} indicates the level is not measured by the
item.
}

\code{pie_pilot_data} is a \link[tibble:tibble-package]{tibble} containing PIE
response data from the pilot administration with 2,370 rows and 19 variables.
\itemize{
\item \code{student}: Respondent identifier.
\item \code{time}: The time point of the assessment (\code{baseline}, \code{midway},
\code{end_of_unit}).
\item \code{00592}-\code{63088}: Dichotomous item responses to the PIE items administered
during the pilot for the 5.RA.A.1b learning pathway.
}

\code{pie_pilot_qmatrix} is a \link[tibble:tibble-package]{tibble} that
identifies which levels of the learning pathway are measured by each PIE
item. The PIE pilot utilized 17 items to measure the 3 levels in the
5.RA.A.1b pathway. The \code{pie_pilot_qmatrix} correspondingly is
made up of 17 rows and 4 variables.
\itemize{
\item \code{task}: Item identifier, corresponds to \code{00592}-\code{63088} in
\code{pie_pilot_data}.
\item \code{L1}, \code{L2}, and \code{L2}: Dichotomous indicator for whether or not the level is
measured by each item. A value of \code{1} indicates the level is measured by
the item and a value of \code{0} indicates the level is not measured by the
item.
}
}
\usage{
pie_ft_data

pie_ft_qmatrix

pie_pilot_data

pie_pilot_qmatrix
}
\description{
PIE was a proof-of-concept assessment partnership between Accessible
Teaching, Learning, and Assessment Systems (\href{https://atlas.ku.edu}{ATLAS})
at the University of Kansas and the Missouri Department of Elementary and
Secondary Education (\href{https://dese.mo.gov/}{DESE}). The goal of the PIE
project was to demonstrate the utility of through-year instructionally
embedded assessments for both informing instructional
decisions and reporting summative results. A learning pathway was developed
for each learning standard so that teachers could identify precisely where
students are in their learning journey. There are two data sets containing
student responses to a learning pathway aligned to the 5.RA.A.1b learning
standard. The first data set includes responses from the Spring 2024 field
test, and the second includes responses from the 2024--2025 pilot
administration.
}
\details{
\subsection{Learning Pathway}{

The 5.RA.A.1b is a grade 5 mathematics standard in the Relationships and
Algebraic Thinking (RA) domain. Specifically, this standard comes from
Cluster A (Represent and analyze patterns and relationships), and represents
Expectation 1 (Investigate the relationship between two numeric patterns),
sub-expectation B (Translate two numeric patterns into two sets of ordered
pairs).

The learning pathway that was developed for 5.RA.A.1b includes three vertical
levels that build toward the intended learning target. Level 1 (L1) includes
emerging concepts and skills, Level 2 (L2) includes skills approaching the
learning target, and Level 3 (L3) represents the learning target and is
directly aligned to the learning standard. For this learning pathway, the
levels are:
\itemize{
\item Level 1: Recognize the order of elements in a repeating pattern.
\item Level 2: Organize two numeric patterns in a table.
\item Level 3: Translate two numeric patterns into ordered pairs.
}

For additional information on the development of the learning pathways, see
Kim et al. (2024).
}

\subsection{Assessment Design}{

The PIE field test was administered as a fixed-form assessment during Spring
2024. During the field test, each student was assessed on all three levels
within the learning pathway at a single point in time. Following the field
test, the test development team reviewed item data and selected items for
promotion to the pilot item pool. \code{pie_ft_data} includes only items that were
promoted for use during the pilot administration. For more information on the
PIE field test, see ATLAS (2025a).

During the pilot administration, teachers chose when to assess each of the
learning standards. PIE assessments were intended to be administered at three
points during the instructional cycle. The first assessment (\code{baseline}) was
administered before the instructional cycle. The baseline assessment measured
only Level 1 skills, allowing teachers to evaluate whether their students
were ready for instruction on the standard. After some instruction had
occurred, teachers administered the second assessment (\code{midway}). The midway
assessment measured Level 2 skills for all students and additionally
re-assessed Level 1 skills for students who did not demonstrate mastery of
those skills on the baseline assessment. Teachers can then observe where
their students are at, evaluate the progress they have made, and plan
additional targeted instruction. The final assessment was administered at the
end of the instructional unit (\code{end_of_unit}). This assessment included items
measuring Level 3 skills for all students, as well as Level 2 skills for
students who did not demonstrate mastery of those skills on the midway
assessment. Teachers could again use this information to monitor their
students' progress and make decisions about next instructional steps.
In total, the three pathway levels were assessed over three timepoints. For
additional information on the pilot design and administration, see ATLAS
(2025b).

Notably, there is not a one-to-one correspondence between items included in
the field test and pilot data sets. There are three items that were promoted
following the field test, but were not administered during the pilot. These
items were either held back in the event that other items included in the
pilot pool had to be removed and replaced during the pilot administration, or
were assigned to a fixed-form end-of-year assessment that is not included in
this data. Additionally, there are five items that were administered during
the pilot that were not directly field tested. These items are "twins" of
other items that were field tested. For details on the item twinning
approach, see ATLAS (2025a).
}
}
\references{
Accessible Teaching, Learning, and Assessment Systems. (2025a).
\emph{PIE assessment design and development}. University of Kansas.
\url{https://pie.atlas4learning.org/}

Accessible Teaching, Learning, and Assessment Systems. (2025b).
\emph{PIE pilot study: Design and administration evidence}. University of Kansas.
\url{https://pie.atlas4learning.org/}

Kim, E. M., Nash, B., & Swinburne Romine, R. (2024). \emph{Pathways
for instructionally embedded assessment (PIE): Developing learning pathways
for the PIE assessment system}. University of Kansas; Accessible Teaching,
Learning, and Assessment Systems. \url{https://pie.atlas4learning.org/}
}
\keyword{datasets}
