% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timss03.R
\docType{data}
\name{timss03_data}
\alias{timss03_data}
\alias{timss03}
\alias{timss03_qmatrix}
\title{Trends in international mathematics and science study (TIMSS) assessment
for grade 8 mathematics (2003)}
\format{
\code{timss03_data} is a \link[tibble:tibble-package]{tibble} containing TIMSS
response data with 757 rows and 24 variables.
\itemize{
\item \code{id}: Respondent identifier.
\item \code{M012001}-\code{M022234B}: Dichotomous item responses to the 23 TIMSS items.
}

\code{timss03_qmatrix} is a \link[tibble:tibble-package]{tibble} that
identifies which skills are measured by each TIMSS 2003 item. This data set
contains a subset of data consisting of 23 items measuring 13 skills. The
\code{timss03_qmatrix} correspondingly is made up of 23 rows and 14 variables.
\itemize{
\item \code{item}: Item identifier, corresponds to item response columns in
\code{timss03_data}.
\item Attribute columns: 13 columns, one for each attribute. Each is a
dichotomous indicator for whether or not the skill is measured by each
item. A value of \code{1} indicates the skill is measured by the item and a
value of \code{0} indicates the skill is not measured by the item.
}
}
\usage{
timss03_data

timss03_qmatrix
}
\description{
This is data from the United States sample of the 2003 TIMSS assessment for
grade 8 mathematics. This data contains responses to 23 items from 757
respondents. This data has been used by Skaggs et al. (2016) and Su et al.
(2013) to evaluate the appropriateness of using diagnostic models for
modeling the TIMSS assessment data.
The data set was uploaded to the Item Response Warehouse
(Domingue et al., 2025) and is reformatted here.
}
\details{
The skills correspond to knowledge of:
\itemize{
\item \code{understand_ratio}: Understand concepts of a ratio and a unit rate.
\item \code{use_ratio}: Use ratio and rate reasoning to solve problems.
\item \code{compute_fluently}: Compute fluently with multi-digit numbers.
\item \code{rational_numbers}: Apply and extend understandings of numbers to the
system of rational numbers.
\item \code{algebraic_expressions}: Apply and extend understandings of arithmetic to
algebraic expressions.
\item \code{one_variable_equations}: Solve one-variable equations and inequalities.
\item \code{recognize_proportional_relationships}: Recognize and represent
proportional relationships between quantities.
\item \code{use_proportional_relationships}: Use proportional relationships to solve
multi-step ratio and percent problems.
\item \code{asmd_rational_numbers}: Add, subtract, multiply, and divide rational
numbers.
\item \code{expressions_equations}: Solve problems using numerical and algebraic
expressions and equations.
\item \code{compare_fractions}: Compare two fractions with different numerators and
denominators.
\item \code{multistep_problems}: Solve multi-step problems with whole numbers using
the four operations.
\item \code{equivalent_fractions}: Use equivalent fraction as a strategy to add and
subtract fractions.
}

For more details, see Table 2 of Su et al. (2013).
}
\references{
Domingue, B., Braginsky, M., Caffrey-Maffei, L., Gilbert, J. B.,
Kanopka, K., Kapoor, R., Lee, H., Liu, Y., Nadela, S., Pan, G., Zhang, L.,
Zhang, S., & Frank, M. C. (2025). An introduction to the Item Response
Warehouse (IRW): A resource for enhancing data usage in psychometrics.
\emph{Behavior Research Methods, 57}, Article 276.
\doi{10.3758/s13428-025-02796-y}

Skaggs, G., Wilkins, J. L. M., & Hein, S. F. (2016). Grain size
and parameter recovery with TIMSS and the general diagnostic model.
\emph{International Journal of Testing, 16}(4), 310-330.
\doi{10.1080/15305058.2016.1145683}

Su, Y.-L., Choi, K. M., Lee, W.-C., Choi, T., & McAninch, M.
(2013). \emph{Hierarchical cognitive diagnostic analysis for TIMSS 2003
mathematics} (Research Report No. 35). University of Iowa, Center for
Advanced Studies in Measurement and Assessment.
\url{https://education.uiowa.edu/sites/education.uiowa.edu/files/2022-10/casma-research-report-35.pdf}
}
\keyword{datasets}
