% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcmdata-package.R
\docType{package}
\name{dcmdata-package}
\alias{dcmdata}
\alias{dcmdata-package}
\title{dcmdata: Data Sets for Diagnostic Classification Modeling}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

More about what it does (maybe more than one line). Continuation lines should be indented.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://dcmdata.r-dcm.org}
  \item \url{https://github.com/r-dcm/dcmdata}
  \item \url{https://r-dcm.github.io/dcmdata/}
  \item Report bugs at \url{https://github.com/r-dcm/dcmdata/issues}
}

}
\author{
\strong{Maintainer}: W. Jake Thompson \email{wjakethompson@gmail.com} (\href{https://orcid.org/0000-0001-7339-0300}{ORCID})

Other contributors:
\itemize{
  \item University of Kansas [copyright holder]
  \item Institute of Education Sciences [funder]
}

}
\keyword{internal}
